Implement event pool
