+++ cpp/include/kvikio/detail/event.hpp
@@ -1,0 +1,10 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2026, NVIDIA CORPORATION.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+#pragma once
+
+#include <kvikio/shim/cuda.hpp>
+
+namespace kvikio::detail {
+}
